[noticket] Add legacy support - Updated config to work with older versions - Updated php-cs-fixer - Changed fixer-registration to return defined fixers - Add 'blank_line_before_statement' as style
