+++ .php_cs.dist
@@ -2,8 +2,7 @@
 
 require 'vendor/autoload.php';
 
-return \K10rFixer\Rules\PHP72::create(
+return \K10r\Codestyle\PHP72::create(
     (\PhpCsFixer\Finder::create()
         ->in(__DIR__))
-)
-    ->registerCustomFixers(new \K10rFixer\CustomFixer());
+);
