+++ composer.json
@@ -17,7 +17,7 @@
   ],
   "autoload": {
     "psr-4": {
-      "K10rFixer\\": "src/"
+      "K10r\\": "src/"
     }
   },
   "bin": [
