+++ src/Codestyle/PHP56.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace K10r\Codestyle;
+
+final class PHP56 extends DefaultRules
+{
+    /**
+     * @return array
+     */
+    public static function getRules()
+    {
+        return self::RULES;
+    }
+}
